feat: refactor  paginator
